Repository 타입 Integer에서 Long으로 수정

+++ src/main/java/com/allrounders/goalkeeper/repository/LikesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LikesRepository extends JpaRepository<Likes, Integer> {
+public interface LikesRepository extends JpaRepository<Likes, Long> {
 }
